add calculate test
